requirements.txt
discord.py==2.1.0
PyNaCl==1.5.0